refactor(controller): simplify controller logic by moving UI rendering out
refactor(controller): decouple ControllerUI from BaseConnector
